This Ruby  script should output: [SENDER],[RECEIVER],[FLAGS] - The sender phone number or name (including country code if present) - The receiver phone number or name (including country code if present) - The flags that were used
the regular expression /\[from:(.*?)\] \[to:(.*?)\] \[flags:(.*?)\]/ is designed to extract specific information from a string that follows the pattern of "[from:...], [to:...], [flags:...]". The three capturing groups capture the values between the respective markers, allowing you to extract the specific data enclosed within the brackets.

In summary, this regex is designed to match strings in the format:

[from:<content1>] [to:<content2>] [flags:<content3>]
<content1> will be captured by the first capturing group.
<content2> will be captured by the second capturing group.
<content3> will be captured by the third capturing group.

The non-greedy quantifiers (.*?) ensure that the capturing groups match as little characters as possible, allowing the regex to capture the content between the specific markers.